convert remaining uint64 nonces to uint32

+++ contracts/IRegistry.sol
@@ -123,5 +123,5 @@ interface IRegistry {
      *
      * @return nonce value for handle
      */
-    function resolveHandleToNonce(string calldata handle) external view returns (uint64);
+    function resolveHandleToNonce(string calldata handle) external view returns (uint32);
 }
